refactor: Move printer in its own directory
