+++ crates/rome_formatter/src/printer/mod.rs
@@ -1,0 +1,5 @@
+mod printer;
+mod printer_options;
+
+pub use printer::Printer;
+pub use printer_options::*;
